Made changes to the resource tab and home tab
I fixed up the resource tab to have information on the resource tabs such as the advising center and registrars office and all the catalogs.  I also cleaned up the home page and implemented email services that work with outlook and made a notes section that will save the notes in a text file on the users directory. it still needs some work.